Refactored the `linode_api.objects.linode` module
This module is now a flat file so that class paths are shorter, like
`linode_api.objects.linode.Config` instead of
`linode.objects.linode.config.Config`.
