--- linode_api/objects/linode/__init__.py
@@ -1,7 +1,0 @@
-from .linode import Linode
-from .kernel import Kernel
-from .disk import Disk
-from .stackscript import StackScript
-from .config import Config
-from .backup import Backup
-from .linode_type import Type
